+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.1.0] - 2024-08-24
+### Added
+- Debug Bar feature for enhanced developer insights
+  - Real-time websocket connection/disconnect status
+  - MQTT connection status and last message details
+  - Request duration tracking
+  - Toggle functionality to show/hide the Debug Bar
+
 ## [1.0.0] - 2024-08-19
 ### Added
 - Initial release of MQTT Web Interface